+++ Model/Client/Api/Payment.php
@@ -1,6 +1,5 @@
 <?php
 
-
 namespace Dibs\EasyCheckout\Model\Client\Api;
 
 use Dibs\EasyCheckout\Model\Client\Client;
@@ -15,17 +14,33 @@ use Dibs\EasyCheckout\Model\Client\DTO\RefundPayment;
 use Dibs\EasyCheckout\Model\Client\DTO\UpdatePaymentCart;
 use Dibs\EasyCheckout\Model\Client\DTO\CreatePaymentResponse;
 use Dibs\EasyCheckout\Model\Client\DTO\UpdatePaymentReference;
+use Dibs\EasyCheckout\Model\Client\DTO\ReportingRequest;
+use Dibs\EasyCheckout\Model\Client\DTO\ReportingResponse;
 
-class Payment extends Client
-{
+class Payment extends Client {
+
+    /**
+     * @param ReportingRequest $createPayment
+     * @return ReportingResponse
+     * @throws ClientException
+     */
+	public function reportingApi(ReportingRequest $createPayment) {
+		    try {
+            $response = $this->post("https://reporting.sokoni.it/enquiry", $createPayment);
+        } catch (ClientException $e) {
+            // handle?
+            throw $e;
+        }
+
+        return new ReportingResponse($response);
+    }
 
     /**
      * @param CreatePayment $createPayment
      * @return CreatePaymentResponse
      * @throws ClientException
      */
-    public function createNewPayment(CreatePayment $createPayment)
-    {
+    public function createNewPayment(CreatePayment $createPayment) {
         try {
             $response = $this->post("/v1/payments", $createPayment);
         } catch (ClientException $e) {
@@ -36,22 +51,19 @@ class Payment extends Client
         return new CreatePaymentResponse($response);
     }
 
-
     /**
      * @param UpdatePaymentCart $cart
      * @param $paymentId
      * @return void
      * @throws \Exception
      */
-    public function UpdatePaymentCart(UpdatePaymentCart $cart, $paymentId)
-    {
+    public function UpdatePaymentCart(UpdatePaymentCart $cart, $paymentId) {
         try {
-            $this->put("/v1/payments/".$paymentId."/orderitems", $cart);
+            $this->put("/v1/payments/" . $paymentId . "/orderitems", $cart);
         } catch (ClientException $e) {
             // handle?
             throw $e;
         }
-
     }
 
     /**
@@ -61,15 +73,13 @@ class Payment extends Client
      * @return void
      * @throws ClientException
      */
-    public function UpdatePaymentReference(UpdatePaymentReference $reference, $paymentId, $storeId)
-    {
+    public function UpdatePaymentReference(UpdatePaymentReference $reference, $paymentId, $storeId) {
         try {
-            $this->put("/v1/payments/".$paymentId."/referenceinformation", $reference, $storeId);
+            $this->put("/v1/payments/" . $paymentId . "/referenceinformation", $reference, $storeId);
         } catch (ClientException $e) {
             // handle?
             throw $e;
         }
-
     }
 
     /**
@@ -78,8 +88,7 @@ class Payment extends Client
      * @return GetPaymentResponse
      * @throws ClientException
      */
-    public function getPayment($paymentId, $storeId)
-    {
+    public function getPayment($paymentId, $storeId) {
         try {
             $response = $this->get("/v1/payments/" . $paymentId, $storeId);
         } catch (ClientException $e) {
@@ -90,15 +99,13 @@ class Payment extends Client
         return new GetPaymentResponse($response);
     }
 
-
     /**
      * @param CancelPayment $payment
      * @param string $paymentId
      * @throws ClientException
      * @return void
      */
-    public function cancelPayment(CancelPayment $payment, $paymentId)
-    {
+    public function cancelPayment(CancelPayment $payment, $paymentId) {
         try {
             $this->post("/v1/payments/" . $paymentId . "/cancels", $payment);
         } catch (ClientException $e) {
@@ -113,8 +120,7 @@ class Payment extends Client
      * @throws ClientException
      * @return CreatePaymentChargeResponse
      */
-    public function chargePayment(ChargePayment $payment, $paymentId)
-    {
+    public function chargePayment(ChargePayment $payment, $paymentId) {
         try {
             $response = $this->post("/v1/payments/" . $paymentId . "/charges", $payment);
         } catch (ClientException $e) {
@@ -125,17 +131,15 @@ class Payment extends Client
         return new CreatePaymentChargeResponse($response);
     }
 
-
     /**
      * @param RefundPayment $paymentCharge
      * @param string $chargeId
      * @throws ClientException
      * @return CreateRefundResponse
      */
-    public function refundPayment(RefundPayment $paymentCharge, $chargeId)
-    {
+    public function refundPayment(RefundPayment $paymentCharge, $chargeId) {
         try {
-           $response = $this->post("/v1/charges/" . $chargeId . "/refunds", $paymentCharge);
+            $response = $this->post("/v1/charges/" . $chargeId . "/refunds", $paymentCharge);
         } catch (ClientException $e) {
             // handle?
             throw $e;